services

+++ ScrumMasters.Webshop.DataAccess/Repositories/CategoryRepository.cs
@@ -11,11 +11,12 @@ namespace ScrumMasters.Webshop.DataAccess.Repositories
     public class CategoryRepository : ICategoryRepository
     {
         private readonly MainDbContext context;
-        
+
         public CategoryRepository(MainDbContext ctx)
         {
             context = ctx ?? throw new InvalidDataException("Category Repository Must have a DBContext");
         }
+
         public List<Category> FindAll()
         {
             return context.Categories
@@ -23,37 +24,39 @@ namespace ScrumMasters.Webshop.DataAccess.Repositories
                 {
                     Id = pe.Id,
                     Name = pe.Name,
-                    Products = pe.Product.Select(px=>new Product{Id = px.Id,ProductName = px.ProductName}).ToList()
+                    Products = pe.Product.Select(px => new Product {Id = px.Id, ProductName = px.ProductName}).ToList()
                 })
                 .ToList();
         }
-        
-        
-                public Category GetById(int id)
+
+
+        public Category GetById(int id)
         {
+            if (id == 0) return null;
             return context.Categories
                 .Select(pe => new Category
                 {
                     Id = pe.Id,
                     Name = pe.Name,
-                    Products = pe.Product.Select(px=>new Product{Id = px.Id,ProductName = px.ProductName}).ToList()
+                    Products = pe.Product.Select(px => new Product {Id = px.Id, ProductName = px.ProductName}).ToList()
                 }).FirstOrDefault(category => category.Id == id);
         }
-        
+
         public Category DeleteById(int id)
         {
+            if (id == 0) return null;
             var savedEntity = context.Categories.Remove(new CategoryEntity() {Id = id}).Entity;
             context.SaveChanges();
             return new Category()
             {
                 Id = savedEntity.Id,
                 Name = savedEntity.Name,
-
             };
         }
 
         public Category Update(Category category)
         {
+            if (category == null) return null;
             var pe = context.Update(new CategoryEntity
             {
                 Id = category.Id,
@@ -69,6 +72,7 @@ namespace ScrumMasters.Webshop.DataAccess.Repositories
 
         public Category Create(Category category)
         {
+            if (category == null) return null;
             var entity = new CategoryEntity()
             {
                 Name = category.Name,
